+++ app/openai_whisper/core.py
@@ -1,7 +1,9 @@
 import os
 from io import StringIO
-from threading import Lock
+from threading import Lock, Thread
 from typing import BinaryIO, Union
+import time
+import gc
 
 import torch
 import whisper
@@ -9,13 +11,41 @@ from whisper.utils import ResultWriter, WriteJSON, WriteSRT, WriteTSV, WriteTXT,
 
 model_name = os.getenv("ASR_MODEL", "base")
 model_path = os.getenv("ASR_MODEL_PATH", os.path.join(os.path.expanduser("~"), ".cache", "whisper"))
-
-if torch.cuda.is_available():
-    model = whisper.load_model(model_name, download_root=model_path).cuda()
-else:
-    model = whisper.load_model(model_name, download_root=model_path)
+model = None
 model_lock = Lock()
 
+last_activity_time = time.time()
+idle_timeout = int(os.getenv("IDLE_TIMEOUT", 0))  # default to being disabled
+
+def monitor_idleness():
+    global model
+    if(idle_timeout <= 0): return
+    while True:
+        time.sleep(15)  # check every minute
+        if time.time() - last_activity_time > idle_timeout:
+            with model_lock:
+                release_model()
+                break
+
+def load_model():
+    global model
+
+    if torch.cuda.is_available():
+        model = whisper.load_model(model_name, download_root=model_path).cuda()
+    else:
+        model = whisper.load_model(model_name, download_root=model_path)
+
+    Thread(target=monitor_idleness, daemon=True).start()
+
+load_model()
+
+def release_model():
+    global model
+    del model
+    torch.cuda.empty_cache()
+    gc.collect()
+    model = None
+    print("Model unloaded due to timeout")
 
 def transcribe(
     audio,
@@ -26,6 +56,12 @@ def transcribe(
     word_timestamps: Union[bool, None],
     output,
 ):
+    global last_activity_time
+    last_activity_time = time.time()
+
+    with model_lock:
+        if(model is None): load_model()
+
     options_dict = {"task": task}
     if language:
         options_dict["language"] = language
@@ -44,6 +80,12 @@ def transcribe(
 
 
 def language_detection(audio):
+    global last_activity_time
+    last_activity_time = time.time()
+
+    with model_lock:
+        if(model is None): load_model()
+
     # load audio and pad/trim it to fit 30 seconds
     audio = whisper.pad_or_trim(audio)