+++ app/asr_models/faster_whisper_engine.py
@@ -34,8 +34,6 @@ class FasterWhisperASR(ASRModel):
             word_timestamps: Union[bool, None],
             output,
     ):
-
-        print("faster whisper")
         self.last_activity_time = time.time()
 
         with self.model_lock: